+++ tests/clients/test_secure_http_service_invocation_client.py
@@ -51,7 +51,7 @@ class DaprSecureInvocationHttpClientTests(DaprInvocationHttpClientTests):
         self.server.start()
         settings.DAPR_HTTP_PORT = self.server_port
         settings.DAPR_API_METHOD_INVOCATION_PROTOCOL = 'http'
-        settings.DAPR_HTTP_ENDPOINT = "https://localhost:{}".format(self.server_port)
+        settings.DAPR_HTTP_ENDPOINT = 'https://localhost:{}'.format(self.server_port)
         self.client = DaprClient()
         self.app_id = 'fakeapp'
         self.method_name = 'fakemethod'
@@ -118,13 +118,3 @@ class DaprSecureInvocationHttpClientTests(DaprInvocationHttpClientTests):
         self.server.set_server_delay(1.5)
         with self.assertRaises(TimeoutError):
             new_client.invoke_method(self.app_id, self.method_name, '')
-
-    @patch.object(settings, "DAPR_HTTP_ENDPOINT", None)
-    def test_get_api_url_default(self):
-        client = DaprClient()
-        self.assertEqual(
-            'http://{}:{}/{}'.format(
-                settings.DAPR_RUNTIME_HOST, settings.DAPR_HTTP_PORT, settings.DAPR_API_VERSION
-            ),
-            dapr.clients.http.helpers.get_api_url(),
-        )
